+++ src/server/src/districts/controllers/districts.controller.ts
@@ -15,7 +15,7 @@ import {
   DistrictsImportApiResponse,
   DistrictImportField
 } from "../../../../shared/entities";
-import { FIPS } from "../../../../shared/constants";
+import { FIPS, MAX_IMPORT_ERRORS } from "../../../../shared/constants";
 import { TopologyService } from "../services/topology.service";
 
 import { RegionConfigsService } from "../../region-configs/services/region-configs.service";
@@ -57,11 +57,27 @@ export class DistrictsController {
     if (!(stateFips in FIPS)) {
       throw new InternalServerErrorException();
     }
+
+    const regionCode = FIPS[stateFips];
+    const regionConfig = await this.regionConfigService.findOne({
+      regionCode,
+      hidden: false,
+      archived: false
+    });
+    const geoCollection = regionConfig && (await this.topologyService.get(regionConfig.s3URI));
+    if (!geoCollection) {
+      throw new InternalServerErrorException();
+    }
+    if (!("topology" in geoCollection)) {
+      // Only unarchived regions support imports
+      throw new InternalServerErrorException();
+    }
+
     const blockIdCounts: {
       [blockId: string]: number;
     } = {};
 
-    // Iterate through records again to flag invalid rows
+    // Iterate through records to flag invalid rows
     records.forEach((record, i) => {
       const rowFips = record[0]?.slice(0, 2);
       const blockId = record[0];
@@ -87,50 +103,46 @@ export class DistrictsController {
       }
     });
 
-    const regionCode = FIPS[stateFips];
-    const regionConfig = await this.regionConfigService.findOne({
-      regionCode,
-      hidden: false,
-      archived: false
-    });
-    const geoCollection = regionConfig && (await this.topologyService.get(regionConfig.s3URI));
-    if (!geoCollection) {
-      throw new InternalServerErrorException();
-    }
-    if (!("topology" in geoCollection)) {
-      // Only unarchived regions support imports
-      throw new InternalServerErrorException();
-    }
-
     const unflaggedRows = records.filter((record, i) => !flaggedRows[i]);
     const baseGeoLevel = geoCollection.definition.groups.slice().reverse()[0];
     const baseGeoUnitProperties = geoCollection.topologyProperties[baseGeoLevel];
-
-    const blockToDistricts = Object.fromEntries(
-      unflaggedRows.map(([block, district]) => [block, Number(district)])
-    );
-    const districtsDefinition = geoCollection.importFromCSV(blockToDistricts);
 
     // Find unmatched records
     const allBlockIds: Set<unknown> = new Set(
       // eslint-disable-next-line @typescript-eslint/no-unsafe-return
       baseGeoUnitProperties.map((props: any) => props[baseGeoLevel])
     );
-    records.forEach((record, i) => {
+    const invalidRecords = records.filter((record, i) => {
       if (!allBlockIds.has(record[0]) && !flaggedRows[i]) {
         setFlag(record, i, "BLOCKID", "Invalid block ID");
+        return true;
       }
+      return false;
     });
+
+    // This is a heuristic more than an exact detection method, but it seems sufficient from my testing
+    if (invalidRecords.length > MAX_IMPORT_ERRORS) {
+      return {
+        error: `There were ${invalidRecords.length} invalid block IDs for ${regionCode}, ensure the CSV uploaded is for the Census year ${regionConfig?.census}`
+      };
+    }
+
+    const blockToDistricts = Object.fromEntries(
+      unflaggedRows.map(([block, district]) => [block, Number(district)])
+    );
+    const districtsDefinition = geoCollection.importFromCSV(blockToDistricts);
 
     const maxDistrictId = Math.max(
       ...districtsDefinition.flat(geoCollection.staticMetadata.geoLevels.length)
     );
     const rowFlags = flaggedRows.filter(r => !!r);
+    const numFlags = rowFlags.length;
 
     return {
-      districtsDefinition: districtsDefinition,
-      maxDistrictId: maxDistrictId,
-      rowFlags: rowFlags.length > 0 ? rowFlags : undefined
+      districtsDefinition,
+      maxDistrictId,
+      numFlags: numFlags || undefined,
+      rowFlags: numFlags > 0 ? rowFlags.slice(0, MAX_IMPORT_ERRORS) : undefined
     };
   }
 }